+++ .storybook/main.ts
@@ -2,6 +2,7 @@
 import { createRequire } from "node:module";
 import { dirname, join } from "node:path";
 import type { StorybookConfig } from "@storybook/react-vite";
+import { mergeConfig } from "vite";
 
 const require = createRequire(import.meta.url);
 
@@ -19,10 +20,29 @@ export default {
     getAbsolutePath("@storybook/addon-links"),
     getAbsolutePath("@vueless/storybook-dark-mode"),
   ],
-  framework: getAbsolutePath("@storybook/react-vite"),
+  framework: "@storybook/react-vite",
+  async viteFinal(config) {
+    // Prioritize source files for bundling if available
+    return mergeConfig(config, {
+      resolve: {
+        conditions: ["source"],
+      },
+      optimizeDeps: {
+        exclude: ["node_modules/.cache/storybook"],
+      },
+      css: {
+        preprocessorOptions: {
+          // https://vite.dev/config/shared-options.html#css-preprocessoroptions
+          sass: {
+            api: "modern-compiler", // or "modern", "legacy"
+          },
+        },
+      },
+    });
+  },
   docs: {},
 } as StorybookConfig;
 
 function getAbsolutePath(value: string): any {
-  return dirname(require.resolve(join(value, "package.json")));
+  return dirname(require.resolve(value));
 }