+++ packages/style-system/vite.config.ts
@@ -1,9 +1,14 @@
-import { resolve } from "path";
+import { resolve, dirname } from "node:path";
 import { defineConfig } from "vite";
+import { fileURLToPath } from "node:url";
+
+const __dirname = dirname(fileURLToPath(import.meta.url));
 
 export default defineConfig({
+  root: __dirname,
   build: {
     minify: false,
+    emptyOutDir: true,
     lib: {
       entry: resolve(__dirname, "src/main.sass"),
       name: "@macrostrat/style-system",